Add confirmation when deleting an Article

+++ article/generic_views.py
@@ -21,8 +21,12 @@ def custom_logout(request):
 
 def delete_article_view(request, article_id):
     article = get_object_or_404(Article, id=article_id)
-    article.delete()
-    return HttpResponseRedirect(reverse('my_articles'))
+
+    if request.method == 'POST':
+        article.delete()
+        return redirect('my_articles')
+    else:
+        return render(request, 'article/delete_article.html', {'article': article})
 
 
 @login_required
